+++ src/features/auth/strategy/naver-strategy.ts
@@ -8,6 +8,7 @@ import { GiftogetherExceptions } from 'src/filters/giftogether-exception';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { TokenDto } from '../dto/token.dto';
 import { UserType } from 'src/enums/user-type.enum';
+import { TokenService } from '../token.service';
 
 @Injectable()
 export class NaverStrategy extends PassportStrategy(Strategy, 'naver') {
@@ -15,6 +16,7 @@ export class NaverStrategy extends PassportStrategy(Strategy, 'naver') {
     private readonly configService: ConfigService,
     private readonly authService: AuthService,
     private readonly g2gException: GiftogetherExceptions,
+    private readonly tokenService: TokenService
   ) {
     super({
       clientID: configService.get<string>('NAVER_CLIENT_ID'),
@@ -91,8 +93,8 @@ export class NaverStrategy extends PassportStrategy(Strategy, 'naver') {
       type = "login"
     }
     const tokenDto = new TokenDto();
-    tokenDto.accessToken = await this.authService.createAccessToken(UserType.USER, user.userId);
-    tokenDto.refreshToken = await this.authService.createRefreshToken(user.userId);
+    tokenDto.accessToken = await this.tokenService.createAccessToken(UserType.USER, user.userId);
+    tokenDto.refreshToken = await this.tokenService.createRefreshToken(user.userId);
     done(null, {user, tokenDto, type});
     
   }
